renamed Data.Azure into Azure to not force me to create a multiple projects for Azure Access
