+++ TheLizzards.sln
@@ -13,8 +13,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{893DD448-96BF-4115-94D6-E85E6466F614}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TheLizzards.Data.Azure", "src\TheLizzards.Data.Azure\TheLizzards.Data.Azure.csproj", "{14BA3717-DE31-4495-9267-77ED5AE55812}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TheLizzards.Data", "src\TheLizzards.Data\TheLizzards.Data.csproj", "{5B9A32EC-3601-445E-A5B9-F6DF8BA29AD0}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TheLizzards.Mvc", "src\TheLizzards.Mvc\TheLizzards.Mvc.csproj", "{9DFCA39E-BE89-452E-A8B0-19A54F7AA0BA}"
@@ -29,13 +27,13 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TheLizzards.Mvc.Data.Types"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TheLizzards.Maybe.Tests", "test\TheLizzards.Maybe.Tests\TheLizzards.Maybe.Tests.csproj", "{FEC9CC9D-1C3C-4367-81E1-A59CDD767C54}"
 EndProject
-<<<<<<< HEAD
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TheLizzards.Mvc.Data", "src\TheLizzards.Mvc.Data\TheLizzards.Mvc.Data.csproj", "{FEAE4500-1C95-4D05-A9EF-DE176809ECED}"
-=======
+EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TheLizzards.FileSystem", "src\TheLizzards.FileSystem\TheLizzards.FileSystem.csproj", "{A4BF91DE-6677-43A4-9384-4AA85AF2F75B}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TheLizzards.FileSystem.Tests", "test\TheLizzards.FileSystem.Tests\TheLizzards.FileSystem.Tests.csproj", "{9E5004BE-BAB9-4661-A89C-45FCF1198CF1}"
->>>>>>> 49f2100125e0cb4c18ae920f5cdc7fe423df30c7
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TheLizzards.Azure", "src\TheLizzard.Azure\TheLizzards.Azure.csproj", "{CCB2F5EA-F7EC-47ED-A6F3-7D9CE38A93D7}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -47,18 +45,6 @@ Global
 		Release|x86 = Release|x86
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{14BA3717-DE31-4495-9267-77ED5AE55812}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{14BA3717-DE31-4495-9267-77ED5AE55812}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{14BA3717-DE31-4495-9267-77ED5AE55812}.Debug|x64.ActiveCfg = Debug|x64
-		{14BA3717-DE31-4495-9267-77ED5AE55812}.Debug|x64.Build.0 = Debug|x64
-		{14BA3717-DE31-4495-9267-77ED5AE55812}.Debug|x86.ActiveCfg = Debug|x86
-		{14BA3717-DE31-4495-9267-77ED5AE55812}.Debug|x86.Build.0 = Debug|x86
-		{14BA3717-DE31-4495-9267-77ED5AE55812}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{14BA3717-DE31-4495-9267-77ED5AE55812}.Release|Any CPU.Build.0 = Release|Any CPU
-		{14BA3717-DE31-4495-9267-77ED5AE55812}.Release|x64.ActiveCfg = Release|x64
-		{14BA3717-DE31-4495-9267-77ED5AE55812}.Release|x64.Build.0 = Release|x64
-		{14BA3717-DE31-4495-9267-77ED5AE55812}.Release|x86.ActiveCfg = Release|x86
-		{14BA3717-DE31-4495-9267-77ED5AE55812}.Release|x86.Build.0 = Release|x86
 		{5B9A32EC-3601-445E-A5B9-F6DF8BA29AD0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{5B9A32EC-3601-445E-A5B9-F6DF8BA29AD0}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{5B9A32EC-3601-445E-A5B9-F6DF8BA29AD0}.Debug|x64.ActiveCfg = Debug|x64
@@ -143,7 +129,6 @@ Global
 		{FEC9CC9D-1C3C-4367-81E1-A59CDD767C54}.Release|x64.Build.0 = Release|x64
 		{FEC9CC9D-1C3C-4367-81E1-A59CDD767C54}.Release|x86.ActiveCfg = Release|x86
 		{FEC9CC9D-1C3C-4367-81E1-A59CDD767C54}.Release|x86.Build.0 = Release|x86
-<<<<<<< HEAD
 		{FEAE4500-1C95-4D05-A9EF-DE176809ECED}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{FEAE4500-1C95-4D05-A9EF-DE176809ECED}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{FEAE4500-1C95-4D05-A9EF-DE176809ECED}.Debug|x64.ActiveCfg = Debug|x64
@@ -156,7 +141,6 @@ Global
 		{FEAE4500-1C95-4D05-A9EF-DE176809ECED}.Release|x64.Build.0 = Release|x64
 		{FEAE4500-1C95-4D05-A9EF-DE176809ECED}.Release|x86.ActiveCfg = Release|x86
 		{FEAE4500-1C95-4D05-A9EF-DE176809ECED}.Release|x86.Build.0 = Release|x86
-=======
 		{A4BF91DE-6677-43A4-9384-4AA85AF2F75B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{A4BF91DE-6677-43A4-9384-4AA85AF2F75B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A4BF91DE-6677-43A4-9384-4AA85AF2F75B}.Debug|x64.ActiveCfg = Debug|x64
@@ -181,13 +165,23 @@ Global
 		{9E5004BE-BAB9-4661-A89C-45FCF1198CF1}.Release|x64.Build.0 = Release|x64
 		{9E5004BE-BAB9-4661-A89C-45FCF1198CF1}.Release|x86.ActiveCfg = Release|x86
 		{9E5004BE-BAB9-4661-A89C-45FCF1198CF1}.Release|x86.Build.0 = Release|x86
->>>>>>> 49f2100125e0cb4c18ae920f5cdc7fe423df30c7
+		{CCB2F5EA-F7EC-47ED-A6F3-7D9CE38A93D7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{CCB2F5EA-F7EC-47ED-A6F3-7D9CE38A93D7}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{CCB2F5EA-F7EC-47ED-A6F3-7D9CE38A93D7}.Debug|x64.ActiveCfg = Debug|x64
+		{CCB2F5EA-F7EC-47ED-A6F3-7D9CE38A93D7}.Debug|x64.Build.0 = Debug|x64
+		{CCB2F5EA-F7EC-47ED-A6F3-7D9CE38A93D7}.Debug|x86.ActiveCfg = Debug|x86
+		{CCB2F5EA-F7EC-47ED-A6F3-7D9CE38A93D7}.Debug|x86.Build.0 = Debug|x86
+		{CCB2F5EA-F7EC-47ED-A6F3-7D9CE38A93D7}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{CCB2F5EA-F7EC-47ED-A6F3-7D9CE38A93D7}.Release|Any CPU.Build.0 = Release|Any CPU
+		{CCB2F5EA-F7EC-47ED-A6F3-7D9CE38A93D7}.Release|x64.ActiveCfg = Release|x64
+		{CCB2F5EA-F7EC-47ED-A6F3-7D9CE38A93D7}.Release|x64.Build.0 = Release|x64
+		{CCB2F5EA-F7EC-47ED-A6F3-7D9CE38A93D7}.Release|x86.ActiveCfg = Release|x86
+		{CCB2F5EA-F7EC-47ED-A6F3-7D9CE38A93D7}.Release|x86.Build.0 = Release|x86
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
-		{14BA3717-DE31-4495-9267-77ED5AE55812} = {893DD448-96BF-4115-94D6-E85E6466F614}
 		{5B9A32EC-3601-445E-A5B9-F6DF8BA29AD0} = {893DD448-96BF-4115-94D6-E85E6466F614}
 		{9DFCA39E-BE89-452E-A8B0-19A54F7AA0BA} = {893DD448-96BF-4115-94D6-E85E6466F614}
 		{25DF3EA4-D99E-4F98-AAFF-8B342E0D8827} = {893DD448-96BF-4115-94D6-E85E6466F614}
@@ -195,11 +189,9 @@ Global
 		{6BD3F5F0-B1DE-433A-A9B2-EE39378BDD31} = {893DD448-96BF-4115-94D6-E85E6466F614}
 		{8219AADD-13A5-4F3E-82FF-6169B96606AE} = {893DD448-96BF-4115-94D6-E85E6466F614}
 		{FEC9CC9D-1C3C-4367-81E1-A59CDD767C54} = {ED1E5032-AA42-45AC-96C3-F21666FD4245}
-<<<<<<< HEAD
 		{FEAE4500-1C95-4D05-A9EF-DE176809ECED} = {893DD448-96BF-4115-94D6-E85E6466F614}
-=======
 		{A4BF91DE-6677-43A4-9384-4AA85AF2F75B} = {893DD448-96BF-4115-94D6-E85E6466F614}
 		{9E5004BE-BAB9-4661-A89C-45FCF1198CF1} = {ED1E5032-AA42-45AC-96C3-F21666FD4245}
->>>>>>> 49f2100125e0cb4c18ae920f5cdc7fe423df30c7
+		{CCB2F5EA-F7EC-47ED-A6F3-7D9CE38A93D7} = {893DD448-96BF-4115-94D6-E85E6466F614}
 	EndGlobalSection
 EndGlobal
