+++ src/TheLizzard.Azure/Data/Entities/AzureDatabaseCreator.cs
@@ -5,7 +5,7 @@ using System.Threading.Tasks;
 using Microsoft.Azure.Documents;
 using Microsoft.Azure.Documents.Client;
 
-namespace TheLizzards.Data.Azure.Entities
+namespace TheLizzards.Azure.Data.Entities
 {
 	internal sealed class AzureDatabaseCreator
 	{
