+++ src/TheLizzard.Azure/Data/Materializer.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
 
-namespace TheLizzards.Data.Azure.Entities
+namespace TheLizzards.Azure.Data
 {
 	public static class Materializer
 	{
